create validation class and fix error of not connecting to database

+++ src/main/java/com/workintech/s18d2/repository/FruitRepository.java
@@ -6,17 +6,17 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface FruitRepository extends JpaRepository<Fruit, Integer> {
+public interface FruitRepository extends JpaRepository<Fruit, Long> {
 
-        @Query(value = "SELECT f.id,f.name,f.price,f.fruit_type FROM s18d2.fruit as f " +
-                        "ORDER BY f.price DESC", nativeQuery = true)
-        List<Fruit> getByPriceDesc();
+    @Query(value = "SELECT f.id,f.name,f.price,f.fruit_type FROM s18d2.fruit as f " +
+            "ORDER BY f.price DESC", nativeQuery = true)
+    List<Fruit> getByPriceDesc();
 
-        @Query(value = "SELECT f.id,f.name,f.price,f.fruit_type FROM s18d2.fruit as f " +
-                        "ORDER BY f.price ASC", nativeQuery = true)
-        List<Fruit> getByPriceAsc();
+    @Query(value = "SELECT f.id,f.name,f.price,f.fruit_type FROM s18d2.fruit as f " +
+            "ORDER BY f.price ASC", nativeQuery = true)
+    List<Fruit> getByPriceAsc();
 
-        @Query("SELECT f FROM Fruit f WHERE f.name ILIKE  %:name%")
-        List<Fruit> searchByName(String name);
+    @Query("SELECT f FROM Fruit f WHERE f.name ILIKE  %:name%")
+    List<Fruit> searchByName(String name);
 
 }
